--- community/function-calling/spring-ai-alibaba-starter-function-calling-githubtool/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-com.alibaba.cloud.ai.functioncalling.githubtool.GithubtoolAutoConfiguration
