+++ community/function-calling/spring-ai-alibaba-starter-function-calling-githubtoolkit/src/main/java/com/alibaba/cloud/ai/functioncalling/githubtoolkit/Response.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.cloud.ai.functioncalling.githubtool;
+package com.alibaba.cloud.ai.functioncalling.githubtoolkit;
 
 import com.fasterxml.jackson.annotation.JsonClassDescription;
