+++ community/function-calling/spring-ai-alibaba-starter-function-calling-githubtoolkit/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+com.alibaba.cloud.ai.functioncalling.githubtoolkit.GithubToolKitAutoConfiguration
